moved config logic to a startup class
